requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
jwt==1.2.0
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
